Remove the cases in if-else that will never happen
y == y_prev can only happen in three ways:
1. y == y_prev and y_next < y       - already one of the elif cases
2. y == y_prev and y_next == y    -  handled by "elif y_next == y"
3. y == y_prev and y_next > y       - handle by "elif y_next > y:"

So, `y == y_prev` as a separate case will never be reached. So, removed it.

Similarly `y_prev < y` will also be never reached as the three cases in which this is possible are already accounted for in elifs before it.