Removed unused variables and unused code (#8)
* Remove unused variables

* Update long description in setup.py

* Fix indentation errors

* Make new_min to have a default value of zero

New_min was being reassigned in the code with zero. So, instead we can make it have a default value of zero.

* Remove the cases in if-else that will never happen

y == y_prev can only happen in three ways:
1. y == y_prev and y_next < y       - already one of the elif cases
2. y == y_prev and y_next == y    -  handled by "elif y_next == y"
3. y == y_prev and y_next > y       - handle by "elif y_next > y:"

So, `y == y_prev` as a separate case will never be reached. So, removed it.

Similarly `y_prev < y` will also be never reached as the three cases in which this is possible are already accounted for in elifs before it.

* Remove the dict type check in _scale_x_values

The dict type checking is already done in asciigraph before calling _scale_x_values.
No need to do anything for it here.

* Compute standard deviation and mean only when needed

We need them only if label is set. So, we can compute them only if it is set.